pyproject.toml
[project]
name = "niyam-agent-starters"
version = "0.1.0"
description = "Open-source agent SDK and starter agents for the Niyam AIOL"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [
    { name = "Niyam-Projects Contributors" },
]
keywords = ["ai", "agents", "government", "enterprise", "sdk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

# Minimal dependencies — only what agents and the local runner need.
# Production dependencies (httpx, model SDKs, enterprise connectors,
# FastAPI, observability) live in the closed-source Niyam AIOL.
dependencies = [
    "typer>=0.12,<1",
    "rich>=13,<14",
    "pyyaml>=6,<7",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "jinja2>=3,<4",
    "python-dotenv>=1,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3,<10",
    "pytest-cov>=5,<6",
    "pytest-asyncio>=1,<2",
    "ruff>=0.6,<1",
    "mypy>=1.11,<2",
    "pre-commit>=3,<4",
    "types-PyYAML>=6,<7",
]

[project.scripts]
niyam = "niyam.runner.cli:app"

[project.urls]
Homepage = "https://github.com/Niyam-Projects/government-agent-starters"
Documentation = "https://github.com/Niyam-Projects/government-agent-starters/tree/main/docs"
Repository = "https://github.com/Niyam-Projects/government-agent-starters"
Issues = "https://github.com/Niyam-Projects/government-agent-starters/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/niyam", "agents"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "agents", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "T20",  # flake8-print
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "PTH",  # flake8-use-pathlib
]
ignore = [
    "S101",   # assert used (fine in tests)
    "T201",   # print found (used in CLI)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
addopts = "--strict-markers -v"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
packages = ["niyam", "agents"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
